+++ src/nvcollection/collection_service.py
@@ -18,8 +18,8 @@ class CollectionService(SubController):
     INI_FILEPATH = '.novx/config'
     SETTINGS = dict(
         last_open='',
-        tree_width='260',
-        window_size='600x300',
+        window_geometry='600x300',
+        right_frame_width=350,
     )
     OPTIONS = {}
     ICON = 'cLogo32'
@@ -39,7 +39,7 @@ class CollectionService(SubController):
         self.configuration = self._mdl.nvService.new_configuration(
             settings=self.SETTINGS,
             options=self.OPTIONS
-            )
+        )
         self.configuration.read(self.iniFile)
         self.prefs = {}
         self.prefs.update(self.configuration.settings)
@@ -82,14 +82,10 @@ class CollectionService(SubController):
                 self.collectionView.focus()
                 return
 
-        __, x, y = self._ui.root.geometry().split('+')
-        offset = 100
-        windowPosition = f'+{int(x)+offset}+{int(y)+offset}'
         self.collectionView = CollectionView(
             self._mdl,
             self._ui,
             self._ctrl,
-            windowPosition,
             self.prefs,
         )
         if self.icon: